add new ingredient to meal

+++ lib/feature/foods/data/local/model/food_local_model.dart
@@ -28,7 +28,7 @@ class FoodLocal{
   List<String> fat;
   List<String> diets;
   List<String> allergies;
-  int servingAmount;
+  double servingAmount;
   String unit;
