+++ lib/feature/foods/data/local/model/international_meal_local_model.dart
@@ -20,7 +20,7 @@ class InternationalMealLocal{
   List<String> fat;
   List<String> diets;
   List<String> allergies;
-  int servingAmount;
+  double servingAmount;
   String unit;
