+++ lib/feature/foods/data/local/model/logged_foods_local_model.dart
@@ -25,7 +25,7 @@ class LoggedFoodsLocal{
   List<String> fat;
   List<String> diets;
   List<String> allergies;
-  int servingAmount;
+  double servingAmount;
   String unit;
   String count;
