+++ lib/feature/foods/data/remote/model/generic_food_remote_model.dart
@@ -18,7 +18,7 @@ class GenericFoodRemote with _$GenericFoodRemote{
     @Default('') String description,
     @Default('') String nationality,
     @Default('') String image,
-    @Default([]) List<int> servingAmount,
+    @Default([]) List<double> servingAmount,
     @Default([]) List<String> unit,
     @Default([]) List<String> ingredients,
     @Default([]) List<List<String>> servingIngredientsCount,
