+++ lib/feature/foods/presentation/screen/edit_logged_food_screen.dart
@@ -682,7 +682,7 @@ class _EditLoggedFoodScreenState extends State<EditLoggedFoodScreen> {
       newFood = newFood.copyWith(
           foodType: FoodType.meal,
           name: _mealNameController.text,
-          servingAmount: int.parse(_totalServingController.text),
+          servingAmount: double.parse(_totalServingController.text),
           unit: _totalUnitController.text,
           recipe: _recipeController.text
       );
