+++ lib/feature/foods/presentation/screen/search_foods_list_ui.dart
@@ -123,7 +123,7 @@ class _SearchFoodsListUiState extends State<SearchFoodsListUi> {
               String carb= foodInformation.carb.toStringAsFixed(foodInformation.carb.truncateToDouble() == foodInformation.carb ? 0 : 2);
               String fat= foodInformation.fat.toStringAsFixed(foodInformation.fat.truncateToDouble() == foodInformation.fat ? 0 : 2);
               bool isFoodAdded= foodInformation.isFoodAdded;
-              int servingQuantity= foodInformation.servingQuantity;
+              double servingQuantity= foodInformation.servingQuantity;
               String foodUnit= foodInformation.foodUnit;
 
               return GestureDetector(
