+++ lib/feature/foods/presentation/screen/view_suggested_food_screen.dart
@@ -548,7 +548,7 @@ class _ViewSuggestedFoodScreenState extends State<ViewSuggestedFoodScreen> {
       newFood = newFood.copyWith(
         foodType: FoodType.meal,
         name: _mealNameController.text,
-        servingAmount: int.parse(_totalServingController.text),
+        servingAmount: double.parse(_totalServingController.text),
         unit: _totalUnitController.text,
         recipe: _recipeController.text
       );
